created the deliver method in the bl service and controller

+++ src/main/java/com/scalux/stockManagement/dtos/DeliverDTO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class DeliverDTO {
     private Long id;
-    private Integer deliveredQuantity;
+    private Long deliveredQuantity;
 }
